+++ src/main/kotlin/io/unthrottled/doki/notification/UpdateNotification.kt
@@ -3,8 +3,7 @@ package io.unthrottled.doki.notification
 import com.intellij.ide.plugins.PluginManagerCore.getPlugin
 import com.intellij.ide.plugins.PluginManagerCore.getPluginOrPlatformByClassName
 import com.intellij.notification.Notification
-import com.intellij.notification.NotificationDisplayType
-import com.intellij.notification.NotificationGroup
+import com.intellij.notification.NotificationGroupManager
 import com.intellij.notification.NotificationListener
 import com.intellij.notification.NotificationType
 import com.intellij.notification.impl.NotificationsManagerImpl
@@ -14,10 +13,10 @@ import com.intellij.openapi.project.ProjectManager
 import com.intellij.openapi.ui.popup.Balloon
 import com.intellij.openapi.wm.WindowManager
 import com.intellij.ui.BalloonLayoutData
-import com.intellij.ui.IconManager
 import com.intellij.ui.awt.RelativePoint
 import io.unthrottled.doki.assets.AssetCategory
 import io.unthrottled.doki.assets.AssetManager
+import io.unthrottled.doki.icon.DokiIcons
 import io.unthrottled.doki.themes.ThemeManager
 import org.intellij.lang.annotations.Language
 import org.jetbrains.annotations.Nls
@@ -32,9 +31,15 @@ private fun buildUpdateMessage(
   """
       What's New?<br>
       <ul>
-        <li>Set your own sticker!</li>
-        <li>More content control.</li>
-        <li>New Settings UI!</li>
+        <li>4 New Themes!!
+        <ul>
+            <li>Astolfo (Dark)</li>
+            <li>Rias (extra dark). <br/> 2 Stickers (a cute one and a <b>cultured</b> one!!)</li>
+            <li>Sakuranomiya Maika (Dark)</li>
+            <li>Ayanami Rei (Dark)</li>
+        </ul>
+        <li>Sticker bug fixes</li>
+        <li>Moved lowest supported build to 2020.3.2</li>
       </ul>
       Please see the <a href="https://github.com/doki-theme/doki-theme-jetbrains/blob/master/changelog/CHANGELOG.md">
       changelog</a> for more details.
@@ -51,18 +56,8 @@ private fun buildUpdateMessage(
 
 object UpdateNotification {
 
-  private val NOTIFICATION_ICON = IconManager.getInstance().getIcon(
-    "/icons/doki/Doki-Doki-Logo.svg",
-    UpdateNotification::class.java
-  )
-
-  private val notificationGroup = NotificationGroup(
-    "Doki Theme Updates",
-    NotificationDisplayType.BALLOON,
-    false,
-    "Doki Theme Updates",
-    NOTIFICATION_ICON
-  )
+  private val notificationGroup = NotificationGroupManager.getInstance()
+    .getNotificationGroup("Doki Theme Updates")
 
   private val defaultListener = NotificationListener.UrlOpeningListener(false)
 
@@ -77,7 +72,7 @@ object UpdateNotification {
       title,
       content,
       listener = listener
-    ).setIcon(NOTIFICATION_ICON)
+    ).setIcon(DokiIcons.General.PLUGIN_LOGO)
     actions.forEach {
       notification.addAction(it)
     }
@@ -124,16 +119,16 @@ object UpdateNotification {
         buildUpdateMessage(
           AssetManager.resolveAssetUrl(
             AssetCategory.MISC,
-            "update_celebration_v3.1.gif"
+            "update_celebration_v4.gif"
           ).orElseGet {
-            "https://doki.assets.unthrottled.io/misc/update_celebration_v3.1"
+            "https://doki.assets.unthrottled.io/misc/update_celebration_v4"
           },
           Dimension(450, 253)
         ),
         NotificationType.INFORMATION
       )
         .setListener(NotificationListener.UrlOpeningListener(false))
-        .setIcon(NOTIFICATION_ICON)
+        .setIcon(DokiIcons.General.PLUGIN_LOGO)
     )
   }