+++ src/main/kotlin/io/unthrottled/doki/promotions/AniMemePluginPromotionRunner.kt
@@ -2,11 +2,13 @@ package io.unthrottled.doki.promotions
 
 import com.intellij.ide.IdeEventQueue
 import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.ProjectManager
 import com.intellij.openapi.wm.WindowManager
 import io.unthrottled.doki.themes.ThemeManager
 import io.unthrottled.doki.util.doOrElse
 import io.unthrottled.doki.util.toOptional
+import java.util.Optional
 import java.util.concurrent.TimeUnit
 import kotlin.random.Random
 
@@ -58,10 +60,7 @@ object AniMemePluginPromotion {
   ) {
     ApplicationManager.getApplication().executeOnPooledThread {
       ThemeManager.instance.currentTheme.ifPresent { dokiTheme ->
-        ProjectManager.getInstance().openProjects
-          .toOptional()
-          .filter { it.isNotEmpty() }
-          .map { it.first() }
+        getFirstProject()
           .map {
             WindowManager.getInstance().suggestParentWindow(
               it
@@ -85,3 +84,8 @@ object AniMemePluginPromotion {
     }
   }
 }
+
+fun getFirstProject(): Optional<Project> = ProjectManager.getInstance().openProjects
+  .toOptional()
+  .filter { it.isNotEmpty() }
+  .map { it.first() }