+++ changelog/CHANGELOG.md
@@ -1,6 +1,39 @@
 Changelog
 ---
 
+# 14.0.0 [New Themes!]
+
+## 4 New Themes
+
+From the Fate series:
+
+- Astolfo (Dark)
+
+From the Highschool DxD series:
+
+- Rias: Onyx (a darker theme)
+  - 2 Stickers:
+    - A Mild One
+    - A Cultured One
+
+From the Blend S series:
+
+- Sakuranomiya Maika (Dark)
+
+From the Neon Genesis Evangelion series:
+
+- Ayanami Rei (Dark)
+
+## Other stuff
+
+- Fixed switching between custom/stock sticker [#363](https://github.com/doki-theme/doki-theme-jetbrains/issues/363) 
+- Removed residue from panels when peeling off stickers [#362](https://github.com/doki-theme/doki-theme-jetbrains/issues/362)
+- Many consistency enhancements you probably won't notice :)
+- Moved lowest supported build to 2020.3.2.
+
+![v14 Girls](https://doki.assets.unthrottled.io/misc/v14_girls.png)
+
+
 # 13.2.2 [Consistency Updates]
 
 - Fixed more of the usability issues [found in this issue](https://github.com/doki-theme/doki-theme-jetbrains/issues/340).