+++ changelog/RELEASE-NOTES.md
@@ -1,9 +1,29 @@
-- You now have the ability to set your own custom sticker.
-  Which will be used for any/all the Doki themes.
-- You can now enable/disable content provided by this plugin:
-  - The Sticker in the corner
-  - The empty editor background image
-  - The editor wallpaper
-- Re-Vamped the settings UI
-- Updated icons so, you can read: Mai's, Zero Two's, and Echidna's event log icon.
-- Consistency updates related [to this issue](https://github.com/doki-theme/doki-theme-jetbrains/issues/340)
+## 4 New Themes
+
+From the Fate series:
+
+- Astolfo (Dark)
+
+From the Highschool DxD series:
+
+- Rias: Onyx (a darker theme)
+  - 2 Stickers:
+    - A Mild One
+    - A Cultured One
+
+From the Blend S series:
+
+- Sakuranomiya Maika (Dark)
+
+From the Neon Genesis Evangelion series:
+
+- Ayanami Rei (Dark)
+
+## Other stuff
+
+- Fixed switching between custom/stock sticker [#363](https://github.com/doki-theme/doki-theme-jetbrains/issues/363)
+- Removed residue from panels when peeling off stickers [#362](https://github.com/doki-theme/doki-theme-jetbrains/issues/362)
+- Many consistency enhancements you probably won't notice :)
+- Moved lowest supported build to 2020.3.2.
+
+![v14 Girls](https://doki.assets.unthrottled.io/misc/v14_girls_smol.png)